requirements_minimal.txt
Flask==2.3.2
Flask-Cors==3.0.10
python-dotenv==1.1.1
gspread==5.12.0
google-auth==2.21.0
pandas==2.2.3
openpyxl==3.1.2
requests==2.31.0
openai==1.59.6
PyMuPDF==1.26.6
nltk==3.9.1
scikit-learn==1.7.1
ipython==8.14.0
ipykernel==6.30.1
